backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-slugify
pydantic
slowapi